Allow doubletap/longpress power to toggle torch (9.0) [1/3]
Thanks to beanstown106 for the initial longpress action calls in
PhoneWindowManager (improved by lineage guys)

If the user enables double tap for camera action, torch action will
automatically switch to long press mode.
If the user enables double tap for torch, double tap for camera
will be disabled accordingly.

Remove the settings pref is the device doesn't have flashlight

A 170ms delay for double tap action is not annoying on single tap
to wake up the device.

Vibrate when toggling torch to let the user know of unwanted toggling
in the pocket (even if vibration is disabled by user)

Allow torch action also on ambient display

Change-Id: I12da044f86c7b625872607529cf8524615cf576b
Signed-off-by: Adin Kwok <adin.kwok@carbonrom.org>