CarbonGestures: Add default navigation mode [1/2]
**disclaimer**
Although my patches prior to this are working (in portrait mode),
after considering how much work + how unusable this would be in
landscape mode, I am halting my work with this. Feel free to take
this and do whatever you want.

Add navigation mode that defaults to Left swipe for back, Up/Down
swipe for Home, and Right swipe for recents, and. It can only be
triggered from the very bottom of the screen.

Change-Id: I34a6a37e38481914e2130d7729a35521743e9ed1